Plain-text datasets, torch loading, steering sweeps and a viz package
A dataset is now a text file with one example per line and the label in the
first column, because the question worth asking of interpretability data --
are the two classes matched, or does one just talk about different things? --
is answered by reading it, and JSONL answers it badly. Indentation marks a
group, and split() keeps a group whole so a minimal pair cannot straddle it.

  core/prompts.py    parse, write and check the format; every error names file:line
  core/torchdata.py  Datasets and DataLoaders that yield text, never token ids
  core/steering.py   a strength sweep, measuring effect and coherence apart
  cli/commands/data.py   check a file before any model loads; convert; generate
  cli/commands/viz.py    render the charts, in a terminal or to disk
  viz/               one module per subject, over a palette keyed by meaning

data.source gains 'prompts' alongside synthetic and jsonl, so a spec points at
one the way it points at anything else.

Also folds in the zip(..., strict=True) fixes the new linter asked for: every
one of them pairs prompts with their own completions, scores or labels, where
a length mismatch would have truncated silently.

+++ src/core/config.py
@@ -98,7 +98,7 @@ class ModelConfig:
             )
         if frac is None:
             frac = self.probe_layer_frac
-        return min(int(round(frac * self.n_layers)), self.n_layers - 1)
+        return min(round(frac * self.n_layers), self.n_layers - 1)
 
     def layers(self, fracs: List[float]) -> List[int]:
         """Resolve several depth fractions at once, dropping duplicates and keeping order"""
